creating gras[ amd vpxe; gemeratpr' refacting code so it isnt in hw any more
